Support Orange Pi 5 GPIO
This replaces the RaspberryPi GPIO code with code for the Orange Pi 5.

This uses a custom gpiozero pin factory, based on its lgpio (gpiod) pin
factory, but generalized to run on any board instead of just raspberry
pi.

Then the RaspberryPiPinIds enum is modified to list the pins available
on an Orange Pi 5, replacing the Raspberry Pi mapping for now.

With this change, all pin functions should work as expected on
an Orange Pi 5.